Product trader relation

+++ src/services/product.service.ts
@@ -3,17 +3,23 @@ import AppError from "../utils/AppError.util.js";
 import { categoryRepository } from "../repositories/category.repository.js";
 import { productRepository } from "../repositories/product.repository.js";
 import { ProductCreateData, ProductUpdateData } from "../types/product.types.js";
+import { traderProfileRepository } from "../repositories/traderProfile.repository.js";
 
 
 export const productService = {
   async create(data: ProductCreateData) {
+    const trader = await traderProfileRepository.findById(data.traderId);
+
+    if (!trader) {
+      throw new AppError("Trader not found", 404);
+    }
     const category = await categoryRepository.findById(data.categoryId);
 
     if (!category) {
       throw new AppError("Category not found", 404);
     }
 
-    return productRepository.create(data);
+    return productRepository.create({ ...data });
   },
 
   async getAll(query: { search?: string; categoryId?: string }) {
@@ -31,8 +37,12 @@ export const productService = {
   },
 
   async update(id: string, data: ProductUpdateData) {
-    await this.getById(id);
-
+    const product=await this.getById(id);
+    
+    if(data.traderId!==product.traderId){
+      throw new AppError("You are not authorized to update this product", 403);
+    }
+    
     if (data.categoryId) {
       const category = await categoryRepository.findById(data.categoryId);
 
@@ -44,8 +54,11 @@ export const productService = {
     return productRepository.update(id, data);
   },
 
-  async delete(id: string) {
-    await this.getById(id);
+  async delete(id: string,traderId:number) {
+    const product=await this.getById(id);
+    if(traderId!==product.traderId){
+      throw new AppError("You are not authorized to delete this product", 403);
+    }
     return productRepository.delete(id);
   },
 };
